refactor: js to ts
